requirements.txt
PyQt5==5.15.9
pygame==2.5.2
pystray==0.19.4
Pillow==10.1.0
requests==2.31.0
selenium==4.15.2
PyJWT==2.8.1
